requirements.txt
elevenlabs==1.9.0
streamlit==1.35.0
python-dotenv==1.0.1
requests==2.32.3
gTTS==2.5.1
openai==1.30.1
pydub==0.25.1
